resolve warnings

+++ src/TypeCompile.cpp
@@ -35,14 +35,14 @@ namespace Corrosive {
 			if (cv <= 0) {
 				ThrowSpecificError(size, "Array cannot be created with negative or zero size");
 			}
-			self->actual_size = cv;
+			self->actual_size = (unsigned int)cv;
 		}
 		else if (v.t == t_u8 || v.t == t_u16 || v.t == t_u32 || v.t == t_u64) {
 			unsigned long long cv = LLVMConstIntGetZExtValue(v.v);
 			if (cv == 0) {
 				ThrowSpecificError(size, "Array cannot be created with zero size");
 			}
-			self->actual_size = cv;
+			self->actual_size = (unsigned int)cv;
 		}
 		else {
 			ThrowSpecificError(size, "Array type must have constant integer size");
